requirements.txt
atproto>=0.0.55
python-dotenv>=1.0.0
schedule>=1.2.0
